feat: Fixed ordering of members and styled loading
