+++ src/features/TeamComponent/TeamComponent.tsx
@@ -1,27 +1,4 @@
-import { IconType } from 'react-icons';
-
-export interface TeamProps {
-  teamName: string;
-  description: string;
-  bgColor: string;
-  borderColor: string;
-  members: Member[];
-}
-
-interface Link {
-  url: string;
-  icon: IconType;
-}
-
-export interface Member {
-  name: string;
-  role: string;
-  about: string;
-  imageUrl: string;
-  links: Link[];
-  teamName: string;
-  socialLinks: string[];
-}
+import { TeamProps } from "./TeamComponent.model"
 
 const TeamComponent = ({ teamName, description, bgColor, borderColor, members }: TeamProps) => {
   return (
